% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cudar-package.R
\docType{package}
\name{cudar-package}
\alias{cudar}
\alias{cudar-package}
\title{cudar: cudar}
\description{
A simple R package that uses Rcpp and NVIDIA's CUDA. This is a learning exercise.
}
\keyword{internal}
