% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cudar.R
\name{getDeviceProps}
\alias{getDeviceProps}
\title{getDeviceProps}
\usage{
getDeviceProps(id = 0)
}
\arguments{
\item{id}{ID of the CUDA device for which properties should be returned.
Defaults to \code{0} for the first device.}
}
\description{
The \code{getDeviceProps} function returns some properties of the CUDA
device specified by the \code{id} argument.
}
